requirements.txt
requests
feedparser
python-dotenv
python-telegram-bot[job-queue]
schedule
pytz